refactoring app, filteringPage, mainPage and rename pairings component and remove FindYourWine
