+++ src/routes/filteringPage.js
@@ -4,12 +4,94 @@ import Filter from '../components/filteringPage/filter';
 import FilteredList from '../components/filteringPage/filteredList';
 import React from 'react';
 import axios from 'axios';
+import { useLocation } from 'react-router-dom';
 
-function FilteringPage(props) {
-  const [wines, setWines] = React.useState([]);
+function FilteringPage() {
+  const [filteredWines, setFilteredWines] = React.useState([]);
   const [ratingValue, setRatingValue] = React.useState(3);
   const [ratingHover, setRatingHover] = React.useState(-1);
-  const debouncedWinesList = useDebounce(wines, 5000);
+
+  const [countryState, setCountryState] = React.useState({
+    argentina: [false, '아르헨티나'],
+    australia: [false, '호주'],
+    chile: [false, '칠레'],
+    france: [false, '프랑스'],
+    germany: [false, '독일'],
+    italy: [false, '이탈리아'],
+    newzealand: [false, '뉴질랜드'],
+    spain: [false, '스페인'],
+    usa: [false, '미국'],
+  });
+
+  const [flavorState, setFlavorState] = React.useState({
+    sweet: [2, 4],
+    acidic: [2, 4],
+    body: [2, 4],
+  });
+
+  const [pairingsState, setPairingsState] = React.useState({
+    beef: [false, '소고기'],
+    pork: [false, '돼지고기'],
+    poultry: [false, '가금류'],
+    fish: [false, '생선'],
+    seafood: [false, '해산물'],
+    pasta: [false, '파스타'],
+    cheese: [false, '치즈'],
+    fruit: [false, '과일'],
+    vagetable: [false, '야채'],
+  });
+
+  const [wineState, setWineState] = React.useState({
+    red: [false, '레드'],
+    white: [false, '화이트'],
+    rose: [false, '로제'],
+    sparkling: [false, '스파클링'],
+  });
+
+  const debouncedCountry = useDebounce(countryState, 500);
+  const debouncedFlavor = useDebounce(flavorState, 500);
+  const debouncedPairings = useDebounce(pairingsState, 500);
+  const debouncedWinesType = useDebounce(wineState, 500);
+  const debouncedRating = useDebounce(ratingValue, 500);
+
+  const location = useLocation();
+
+  console.log(debouncedFlavor);
+
+  let mainPageState = location.state;
+  console.log(mainPageState);
+
+  const selectFlavor = (e, value) => {
+    setFlavorState({ ...flavorState, [e]: value });
+  };
+  // 슬라이더 바로 조절한 와인 맛의 값을 state에 반영시켜주는 함수
+
+  const selectPairings = (e) => {
+    if (pairingsState[e][0] && e) {
+      setPairingsState({ ...pairingsState, [e]: [false, pairingsState[e][1]] });
+    } else if (!pairingsState[e][0] && e) {
+      setPairingsState({ ...pairingsState, [e]: [true, pairingsState[e][1]] });
+    }
+  };
+  // 선택한 음식을 true 혹은 false로 변환시켜주는 함수
+
+  const selectCountries = (e) => {
+    if (countryState[e][0] && e) {
+      setCountryState({ ...countryState, [e]: [false, countryState[e][1]] });
+    } else if (!countryState[e][0] && e) {
+      setCountryState({ ...countryState, [e]: [true, countryState[e][1]] });
+    }
+  };
+  // 선택한 국가를 true 혹은 false로 변환시켜주는 함수
+
+  const selectWines = (e) => {
+    if (wineState[e][0] && e) {
+      setWineState({ ...wineState, [e]: [false, wineState[e][1]] });
+    } else if (!wineState[e][0] && e) {
+      setWineState({ ...wineState, [e]: [true, wineState[e][1]] });
+    }
+  };
+  // 선택한 와인 종류를 true 혹은 false로 변환시켜주는 함수
 
   const convertToArr = (obj) => {
     let result = [];
@@ -18,66 +100,68 @@ function FilteringPage(props) {
     }
     return result;
   };
-  // 사용자가 클릭해서 true로 변한 값을 배열로 변환해주는 함수, API 요청을 보내기 위해 사용됨
+  // 사용자가 클릭해서 true로 변한 값의 한글 이름을 배열로 변환해주는 함수
+  // API 요청을 보내기 위해 사용됨
 
-  const handleChange = (event, newValue) => {
+  const handleRatingChange = (event, newValue) => {
     setRatingValue(newValue);
   };
 
-  const handleChangeHover = (event, newHover) => {
+  const handleRatingChangeHover = (event, newHover) => {
     setRatingHover(newHover);
   };
 
-  React.useEffect(
-    () => {
-      const getFilterdList = async () => {
-        try {
-          const res = await axios.get('http://54.180.150.63:3000/wine', {
-            params: {
-              sweet_min: props.flavorState.sweet[0],
-              sweet_max: props.flavorState.sweet[1],
-              acidic_min: props.flavorState.acidic[0],
-              acidic_max: props.flavorState.acidic[1],
-              body_min: props.flavorState.body[0],
-              body_max: props.flavorState.body[1],
-              type: convertToArr(props.wineState),
-              country: convertToArr(props.countryState),
-              food: convertToArr(props.pairingsState),
-              rating: ratingValue,
-            },
-          });
-          setWines(res.data);
-        } catch (error) {
-          console.error(error);
-        }
-      };
-      getFilterdList();
-    },
-    [
-      // debouncedWinesList
-    ],
-  );
+  // React.useEffect(() => {
+  //   setPairingsState(mainPageState.pairingsState);
+  //   setFlavorState(mainPageState.flavorState);
+  // }, []);
 
-  console.log(wines);
-  console.log(debouncedWinesList);
+  React.useEffect(() => {
+    const getFilterdList = async () => {
+      try {
+        const res = await axios.get('http://54.180.150.63:3000/wine', {
+          params: {
+            sweet_min: flavorState.sweet[0],
+            sweet_max: flavorState.sweet[1],
+            acidic_min: flavorState.acidic[0],
+            acidic_max: flavorState.acidic[1],
+            body_min: flavorState.body[0],
+            body_max: flavorState.body[1],
+            type: convertToArr(wineState),
+            country: convertToArr(countryState),
+            food: convertToArr(pairingsState),
+            rating: ratingValue,
+          },
+        });
+        console.log(res.data);
+        setFilteredWines(res.data);
+      } catch (error) {
+        console.error(error);
+      }
+    };
+    // API에 현재 state 값들을 params에 담아 보낸다
+    getFilterdList();
+  }, [debouncedWinesType, debouncedCountry, debouncedFlavor, debouncedPairings, debouncedRating]);
+
+  console.log(filteredWines);
 
   return (
     <Grid container direction="row" justify="center" alignItems="stretch">
       <Filter
-        selectFlavor={props.selectFlavor}
-        selectPairings={props.selectPairings}
-        selectCountries={props.selectCountries}
-        selectWines={props.selectWines}
-        flavorState={props.flavorState}
-        pairingsState={props.pairingsState}
-        countryState={props.countryState}
-        wineState={props.wineState}
+        selectFlavor={selectFlavor}
+        selectPairings={selectPairings}
+        selectCountries={selectCountries}
+        selectWines={selectWines}
+        flavorState={flavorState}
+        pairingsState={pairingsState}
+        countryState={countryState}
+        wineState={wineState}
         ratingHover={ratingHover}
         ratingValue={ratingValue}
-        handleChange={handleChange}
-        handleChangeHover={handleChangeHover}
+        handleRatingChange={handleRatingChange}
+        handleRatingChangeHover={handleRatingChangeHover}
       />
-      <FilteredList wines={wines} />
+      <FilteredList filteredWines={filteredWines} />
     </Grid>
   );
 }
